Split uncovered bytes at block references
gAddrList is a list of all possible address of resource data or directory lists.
gDirList only includes addresses of some resource data or directory lists.
Uncovered bytes will be split at any of these addresses to make it easier to parse the uncovered bytes when they are referenced by a resource.